minor fixes

+++ include/bayes_filter/SensorModel.h
@@ -17,11 +17,11 @@ public:
 	double _pi = 3.14159;
 	double _zhit = 0.8;
 
-	std::vector<std::pair<int,int>> occupied_;
+	std::vector<std::pair<float, float> > occupied_;
 	// SensorModel(nav_msgs::OccupancyGrid map_);
 	SensorModel();
 	void mapUpdate(nav_msgs::OccupancyGrid map_);
-	std::vector<std::pair<int,int>>  _find_objects(nav_msgs::OccupancyGrid map);
+	std::vector<std::pair<float, float> >  _find_objects(nav_msgs::OccupancyGrid map);
 	double _findMinDistance(double x, double y);
 	double gaussianDistribution(double mean, double standardDeviation, double x);
 	void measurementModel(std::vector<Particles> &samples, sensor_msgs::LaserScan scan, nav_msgs::OccupancyGrid map);
